feat(v4v): add value for value setup dialog and manager component
Implement V4V setup dialog and manager component to allow users to configure sales sharing with the community. Includes:
- New V4VSetupDialog component
- V4VManager component with recipient management
- Integration with product publishing flow
- Percentage sliders and visualizations
